Discover documentation page category in parser
