+++ packages/framework/tests/Feature/SourceFileParserTest.php
@@ -3,6 +3,7 @@
 namespace Hyde\Framework\Testing\Feature;
 
 use Hyde\Framework\Actions\SourceFileParser;
+use Hyde\Framework\Hyde;
 use Hyde\Framework\Models\Pages\BladePage;
 use Hyde\Framework\Models\Pages\DocumentationPage;
 use Hyde\Framework\Models\Pages\MarkdownPage;
@@ -81,5 +82,27 @@ class SourceFileParserTest extends TestCase
         $page = MarkdownPage::parse('foo-bar');
 
         $this->assertEquals('Foo Bar', $page->title);
+    }
+
+
+    public function test_documentation_page_parser_can_get_category_from_front_matter()
+    {
+        $this->markdown('_docs/foo.md', '# Foo Bar', ['category' => 'foo']);
+
+        $page = DocumentationPage::parse('foo');
+        $this->assertEquals('foo', $page->category);
+    }
+
+    public function test_documentation_page_parser_can_get_category_automatically_from_nested_page()
+    {
+        mkdir(Hyde::path('_docs/foo'));
+        touch(Hyde::path('_docs/foo/bar.md'));
+
+        /** @var DocumentationPage $page */
+        $page = DocumentationPage::parse('foo/bar');
+        $this->assertEquals('foo', $page->category);
+
+        unlink(Hyde::path('_docs/foo/bar.md'));
+        rmdir(Hyde::path('_docs/foo'));
     }
 }
